+++ src/@components/IsometricPath/IsometricPath.ts
@@ -1,4 +1,5 @@
 import { Command, CommandPoint } from './types';
+import { LineCap, LineJoin, StrokeLinecap, StrokeLinejoin } from '@types';
 import { SVG_NAMESPACE, SVG_ELEMENTS } from '@constants';
 import { Graphic, GraphicProps } from '@components/Graphic';
 import { addSVGProperties, getSVGPath } from '@utils';
@@ -11,14 +12,15 @@ export class IsometricPath extends Graphic {
         super(props);
         this.commands = [];
         this.path = document.createElementNS(SVG_NAMESPACE, SVG_ELEMENTS.path);
-
+        
         addSVGProperties(this.path, {
-            fill: this.fillColor,
-            stroke: this.strokeColor,
+            'fill': this.fillColor,
             'fill-opacity': `${this.fillOpacity}`,
+            'stroke': this.strokeColor,
             'stroke-dasharray': this.strokeDasharray.join(','),
             'stroke-linecap': this.strokeLinecap,
             'stroke-linejoin': this.strokeLinejoin,
+            'stroke-opacity': `${this.strokeOpacity}`,
             'stroke-width': `${this.strokeWidth}`
         });
 
@@ -27,6 +29,46 @@ export class IsometricPath extends Graphic {
     private commands: CommandPoint[];
     private path: SVGPathElement;
     private commandsReg = /(M|L)\s*(\d+\.?\d*|\.\d+)\s*,\s*(\d+\.?\d*|\.\d+)\s*,\s*(\d+\.?\d*|\.\d+)/g;
+
+    protected setFillColor(value: string): void {
+        super.setFillColor(value);
+        addSVGProperties(this.path, { 'fill': this.fillColor });
+    }
+
+    protected setFillOpacity(value: number): void {
+        super.setFillOpacity(value);
+        addSVGProperties(this.path, { 'fill-opacity': `${this.fillOpacity}` });
+    }
+
+    protected setStrokeColor(value: string): void {
+        super.setStrokeColor(value);
+        addSVGProperties(this.path, { 'stroke': this.strokeColor });
+    }
+
+    protected setStrokeDasharray(value: number[]): void {
+        super.setStrokeDashArray(value);
+        addSVGProperties(this.path, { 'stroke-dasharray': this.strokeDasharray.join(',') });
+    }
+
+    protected setStrokeLinecap(value: StrokeLinecap): void {
+        super.setStrokeLinecap(value);
+        addSVGProperties(this.path, { 'stroke-linecap': this.strokeLinecap });
+    }
+
+    protected setStrokeLinejoin(value: StrokeLinejoin): void {
+        super.setStrokeLinejoin(value);
+        addSVGProperties(this.path, { 'stroke-linejoin': this.strokeLinejoin });
+    }
+
+    protected setStrokeOpacity(value: number): void {
+        super.setStrokeOpacity(value);
+        addSVGProperties(this.path, { 'stroke-opacity': `${this.strokeOpacity}` });
+    }
+
+    protected setStrokeWidth(value: number): void {
+        super.setStrokeWidth(value);
+        addSVGProperties(this.path, { 'stroke-width': `${this.strokeWidth}` });
+    }
 
     public update(): void {
         if (this.path.parentNode) {