+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## [1.1.0 - 1.1.1] - 2019-12-20
+## [1.3.0] - 2019-12-20
+
+- IsometricCanvas setter properties now change the SVG element
+- IsometricPath setter properties now change the SVG path element
+- New methods in the IsometricCanvas class (removePath, removePathByIndex, and clear)
+- Fix a bug in the IsometricPath class about missing strokeOpacity property
+
+## [1.1.0 - 1.1.1] - 2019-12-19
 
 - Refactored how to share the canvas properties among shapes using a global class
 - Removed some internal public methods from the IsometricPath class